backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
openai>=1.30.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.7
httpx==0.27.2  # 添加这一行，使用兼容的 httpx 版本
chromadb==0.4.22
sentence-transformers==3.0.0
paddleocr==3.4.0
Pillow==12.1.1